LinkedLists.jl Documentation

Overview

This module contains the structure for (Single) Linked Lists. Additionally, public constructors as well as overloaded functions to work with the iterator protocol, the show method, and the functions push! and pop!.

Approach

We use an internal linked List, ILinkedList, that is Abstract with two subtypes: Cons and Nil. Cons{T} is a container of value, while Nil marks the end of the list.

Notice that the Nil{T} struct has no fields and is NOT mutable. This means that we get a singleton instance which we can compare against to test for the end of a given linked list.

NOTE: Nil{T} is non-mutable while Cons{T} is mutable even though they are subtypes of the abstract type ILinkedList.

Then a LinkedList struct is defined which has a head field that is the internal linked list, ILinkedList.

We use type stability, the singleton nature of Nil{T}, multiple dispatch, and Base operator overloading to create a linked list that is performant with minimal code.

Types

LinkedLists.LinkedList — Type

LinkedList{T}

Fields

head::ILinkedList{T} – An internal linked list.

Outer LinkedList Constructors

LinkedLists.LinkedList — Method

LinkedList(::Type{T})

Creates an empty LinkedList of type T.

Arguments

::Type{T} – The Type of the LinkedList.

Return

Empty LinkedList of type T.

LinkedList(l::LinkedList{T})

Creates a LinkedList which shares the internal list of another LinkedList.

Note: This means that using push! or pop! on one LinkedList will effect the other LinkedList.

Note: If ys = LinkedList(xs) where xs is a LinkedList, then if one pop!'s xs to exhaustion, ys will have one element. That is, pop! will leave one element left in ys.

Arguments

l::LinkedList{T} – Shared LinkedList.

Return

LinkedList of type T.

LinkedList(v::T)

Creates a LinkedList containing a single element, v, of type T.

Arguments

v::T – The sole element of this LinkedList.

Return

LinkedList of type T.

LinkedList(v::Vector{T})

Creates a LinkedList containing the elements from a Vector of type T. The elements of the resulting LinkedList will be in the same order as the vector.

Arguments

v::Vector{T} – The elements of this LinkedList.

Return

LinkedList of type T.

Index

LinkedLists.LinkedList

LinkedLists.LinkedList